driver: use u64 for driver flags
To support mesh mode, use u64 to fit all necessary driver flags.

Signed-hostap: Bob Copeland <me@bobcopeland.com>

--
v5: Added ULL suffix to flag constants, made the format flags more
    politically correct.
v6: Updated email addresses, added signed-hostap

+++ wpa_supplicant/wpa_supplicant_i.h
@@ -561,7 +561,7 @@ struct wpa_supplicant {
 	int scan_for_connection; /* whether the scan request was triggered for
 				  * finding a connection */
 
-	unsigned int drv_flags;
+	u64 drv_flags;
 	unsigned int drv_enc;
 
 	/*
